requirements.txt
# === Core Python and Deep Learning ===
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
torch-geometric>=2.5.2
torchmetrics>=1.3.0
lightning>=2.3.0
transformers>=4.44.0
accelerate>=0.31.0
einops>=0.7.0
timm>=1.0.3

# === Hugging Face and Model Access ===
huggingface_hub>=0.22.2
datasets>=2.20.0
tokenizers>=0.19.1
sentencepiece>=0.2.0

# === Scientific Stack ===
numpy>=1.26.0
scipy>=1.13.1
pandas>=2.2.2
scikit-learn>=1.5.0
statsmodels>=0.14.2
h5py>=3.11.0
anndata>=0.10.5
scanpy>=1.10.2
mudata>=0.3.0

# === Plotting, Visualization, and QC ===
matplotlib>=3.9.0
seaborn>=0.13.2
umap-learn>=0.5.6
plotly>=5.22.0
networkx>=3.3
graphviz>=0.20.3
pyvis>=0.3.2

# === NLP / LLM Utilities (for agent layer and text reasoning) ===
openai>=1.36.0
langchain>=0.2.10
langchain-core>=0.2.24
tiktoken>=0.7.0

# === Bioinformatics and Multi-Omics Utilities ===
biopython>=1.83
pysam>=0.22.0
pybedtools>=0.10.0
scanorama>=1.7.3
decoupler>=1.7.0
gseapy>=1.1.2
goatools>=1.4.14

# === Data I/O and File Management ===
pyarrow>=16.1.0
fastparquet>=2024.2.0
joblib>=1.4.2
tqdm>=4.66.3
requests>=2.32.3
python-dotenv>=1.0.1
pydantic>=2.8.2

# === Machine Learning / Training Enhancements ===
xgboost>=2.1.1
optuna>=3.6.1
wandb>=0.17.5
tensorboard>=2.16.2

# === Image and Spatial Omics (for OmiCLIP, KRONOS, etc.) ===
opencv-python-headless>=4.10.0.84
Pillow>=10.3.0
albumentations>=1.4.8
scikit-image>=0.24.0
tifffile>=2024.7.2

# === Graph Reasoning and Path Planning (Agent Layer) ===
networkx>=3.3
rdflib>=7.0.0
igraph>=0.11.6

# === Notebook / Development Utilities ===
jupyterlab>=4.2.3
ipykernel>=6.29.5
ipywidgets>=8.1.3
notebook>=7.2.1
nbformat>=5.10.4
black>=24.8.0
ruff>=0.5
